+++ tcpClient/getAllStreamEvents.js
@@ -8,7 +8,7 @@ var debug = require('debug')('geteventstore:getAllStreamEvents'),
 var baseErr = 'Get All Stream Events - ';
 
 module.exports = function(config) {
-    return function(streamName, chunkSize) {
+    return function(streamName, chunkSize, startEvent) {
         return q.Promise(function(resolve, reject) {
             assert(streamName, baseErr + 'Stream Name not provided');
 
@@ -32,7 +32,7 @@ module.exports = function(config) {
                     }
                 });
             }
-            getNextChunk(0);
+            getNextChunk(startEvent || 0);
         });
     };
 };